load live channels, infinite scroll

+++ src/twitch/utils.rs
@@ -2,7 +2,7 @@ pub struct TwitchUtils;
 
 impl TwitchUtils {
 
-    pub fn thumbnail_sizer(url: &str, width: u16, height: u16) -> String {
+    pub fn thumbnail_sizer(url: &str, width: i32, height: i32) -> String {
         url.replace("{width}", &width.to_string()).replace("{height}", &height.to_string())
     }
